requirements.txt
tensorflow
opencv-python
ultralytics
streamlit
pandas
numpy
pillow